+++ packages/devui-vue/devui/form/src/components/directive/d-validate.ts
@@ -1,6 +1,6 @@
 import { VNode, DirectiveBinding, h, render, nextTick } from 'vue';
 import { debounce } from 'lodash-es';
-import { EventBus, transformCamelToDash } from '../../util';
+import { EventBus, transformCamelToDash } from '../../utils';
 import useValidate from '../../composables/use-validate';
 import dPopover from '../../../../popover/src/popover';
 import { DFormValidateSubmitData, positionType } from '../../form-types';
